Fix 64-bit build failure.
This function takes an `unsigned int`, not a `size_t`.

Has been broken for about a day:
https://android-review.googlesource.com/#/c/216103/

TreeHugger didn't catch this because there's no 64-bit checkbuild
TreeHugger target (I'm trying to get this fixed, but there's concern
about additional load). The continuous build hasn't caught this
either because the 64-bit targets in aosp are not checkbuilds (aside
from mips64, which is still broken from the NDK update).

Change-Id: Ia30ebb02f33030ff8ac282f46e859a58b7fd041d