fifo: add index references with caching and deferred operations
The new classes will eventually permit a set of related atomic operations,
wakes, and waits to be grouped together into a single sequence point,
and will allow use of event flags to wait for the "or" of multiple FIFOs.

This is one of a series of CLs to isolate the dependencies.

Test: not ready yet for testing
Change-Id: I2055b7a35721222cd914973cd210f5c0dca7d4ef