Make inlines static
Compiling the non-static inlines with -Oz results in missing symbols,
because clang chooses not to inline the functions but also doesn't
emit a definition because they are not static or extern.  They are
only used in this .c file, so make them static.

Test: builds with -Oz
Bug: 62549703
Change-Id: Ibb78ad982c2911e151f0dd6b01eb6164fe6e8e63
Merged-In: Ibb78ad982c2911e151f0dd6b01eb6164fe6e8e63