Audio V4: Update audio.h to V4.0
Exported the audio/common/4.0/types.h to audio-base.h.

As all system only values are removed in 4.0 but the headers
are still shared between vendor and system,
add those values back protected by the macro
AUDIO_NO_SYSTEM_DECLARATIONS.

This macro is defined for vendor libraries.

audio.h used to have *_CNT and *_MAX values in its enum
mostly for enum iteration.

This was not compatible with Treble as it denied any enum
inheritance.

*_CNT and *_MAX are now implemented with constants that can
be updated and are not part of the HIDL API.

Bug: 38184704
Test: compile
Change-Id: I5f99ccc9ce262b3addb15cd5034a885c9ab4cedf
Signed-off-by: Kevin Rocard <krocard@google.com>