Camera: Only obsessively validate camera_metadata on eng builds
Currently, an assert(validate_camera_metadata_structure) is present
at the end of most camera_metadata functions; and asserts are enabled
on both -eng and -userdebug.  This causes large amounts of validate
calls to be done on builds that are used for perf evaluation and dogfood.

Restrict assert enabling to eng builds only.

Test: atest
Bug: 136675319
Change-Id: Ie682bed722cd85f8d495fc5a32fea3f48e41abb5