Stop using audio device type as bit mask for helper functions.
This CL mainly change the helper functions for audio device types which
expecting a single audio device type. Instead of using audio device
types as a bit mask to check if device is a certain kind of device,
using a binary search to do it.

Bug: 135621476
Test: atest systemaudio_tests
Change-Id: Ie8565959af96b09e96ff8ee641876a5e9dc919ea
Merged-In: Ie8565959af96b09e96ff8ee641876a5e9dc919ea