_Generic is only in C11, not in C++
So use overloading instead for C++.

References:
 - https://stackoverflow.com/questions/42496875/why-cant-i-use-generic-in-c-code
 - https://en.cppreference.com/w/c/language/generic

This also exposed an error in one of the metadata test cases: a call to
audio_metadata_get() was unintentionally using the default version, which
caused ASSERT_EQ(nullptr, dataVal) to trivially pass because dataVal had
a stale nullptr.  Fixed test case by adding an &, and pre-initializing
dataVal to a non-nullptr.

Test: metadata_tests
Change-Id: I80a4d4ff70d26933fda18f69e014f9a30934ad49