Downgrade incorrect "E"rror report
In a boot and health check automated test for coral-userdebug, we
get this "E"rror over 1,000 times.  It seems clearly not to
rise to the level of an Error as specified in
https://source.android.com/setup/contribute/code-style#log-sparingly

So we downgrade this to "W"arn level, to prevent "crying wolf"
issues with logcat.

This file has many more uses of ALOGE which seem pretty
questionable, but we start with this clearly egregious one.

Test: TreeHugger
Bug: 163564488
Change-Id: I852fd3b0bba56b0cf0f3a113566fdd7e6cdaadb8