Add audio_port_v7 to describe audio port.
The audio profile is introduced in audio HAL v7.0. Introduce a new
description for audio port which better matches the model of audio
device capabilities. This description is directly compatible with
the definition of audio port in Audio HAL V7.

Within this CL, helper functions for comparing and translating between
audio_port / audio_port_v7 are added.

Test: make, atest systemaudio_tests
Bug: 160352965
Change-Id: I81bb0548f8a24fcb8fc15bc9248fa7a5854a18a7