Add MPEG-H enum.
MPEG-H is a new surround audio codec. Add new enum for MPEG-H in
audio-hal-enums.h so that they can be used by frameworks and HAL.
Although MPEG-H has different profiles and audio level, only low
complexity and baseline with audio level as 3 and 4 is defined. The
formats added are relevant for the current and upcoming commercial
services.

Bug: 178619392
Test: make
Change-Id: I9f72e0c374f1e37902d40864c3c72ec38147b4ab