audio: Specify underlying types for audio enums
When no underlying type is specified, the compiler can
choose it based on the range of declared values. Because
of that, coercing values not declared in the enum definition
can lead to UB. More recent clang versions provide more
stringent checks which can lead to build failures.

Specify underlying types for audio enums as follows:

 - enum types used for lists of values are specified to have
   int32_t type;
 - enum types used for bit masks are specified to have uint32_t
   type.

Bug: 243858318
Test: m
Change-Id: I25aee94868bc034b3581cdd957a3f523d3c35cd3