Support opening stream with a configuration exactly matching the request.
When using alsa proxy to open stream, it may try to open the stream at
maximum sample rate. In this CL, add an extra parameter to indicate the
client wants an exact match for the configuration. This will be needed
when streaming bit-perfectly.

Bug: 262581771
Test: manually
Change-Id: Iaed1b16a22c7fdb1fbe0bcc566d441fab9e004a2
Merged-In: Iaed1b16a22c7fdb1fbe0bcc566d441fab9e004a2