Refactor FdToString
[SAFE CHANGE] New class not in use until flagged in.

Rename existing FdToString for flagging purposes.

Migrate to factory function which returns a writer holding a reference
to the write side of the pipe.

Reader side now non-blocking so timeout is respected if full page is
unavailable.

Reader now closes the read fd when timed out, so writers will fail when
the reader is timed out (SIGPIPE must be suppressed), instead of
blocking indef if the pipe buffer is full.

closeAndGetString now a free function which consumes writer as rvalue
(since write fd should not be referenced after this function and the
function is oneshot).

Add usage documentation + various modern cpp cleanup.

Test: atest fdtostring_tests
Bug: 306283018
Change-Id: Ia6d1d9f4f4dca35d8dcdbe66384025edf0508b39