audio_utils: add missing libsndfile dependency
libsndfile depends on helpers in primitives.c, but that source file
was not part of the cc_library_static sources. Added it to sources.

(This was likely not an issue in the past when libsndfile was used
together with another library that included primitives. This error
would only show up at link time.)

Test: Treehugger
Bug: 308778424
Change-Id: I080108f82414ab8427da967b9d4b5e008a444d7f